feat: re-estructuracion, pequeña implementacion del json
