+++ notas.md
@@ -1,0 +1,5 @@
+- los archivos de la carpeta temp/ son los que no le entendi la funcionalidad/ no le vi la funcionalidad
+
+- la estructura esta algo mas ordenada
+
+- lo que queda por hacer es terminar con la funcionalidad de los botones, mejorar la implementacion del json y seria el pulir el codigo
