+++ temp/componente.py
